reorganized some code
